change unsinged long to uint64_t and size_t

+++ src/string_utils/Trie.cpp
@@ -33,9 +33,9 @@ std::string commonPrefix(const std::string& string1,
     str1 = &string1;
     str2 = &string2;
   }
-  unsigned long size = str1->size();
+  size_t size = str1->size();
   bool flag = false;
-  for (unsigned long i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     if ((*str1)[i] != (*str2)[i]) {
       rtn = str1->substr(0, i);
       flag = true;
